get rid of a few warnings

+++ Source/interfac.cpp
@@ -8,7 +8,7 @@
 DEVILUTION_BEGIN_NAMESPACE
 
 BYTE *sgpBackCel;
-int sgdwProgress;
+DWORD sgdwProgress;
 int progress_id;
 
 /** The colour used for the progress bar as an index into the palette. */
@@ -289,11 +289,11 @@ BOOL IncProgress()
 {
 	interface_msg_pump();
 	sgdwProgress += 23;
-	if ((DWORD)sgdwProgress > 534)
+	if (sgdwProgress > 534)
 		sgdwProgress = 534;
 	if (sgpBackCel != NULL)
 		DrawCutscene();
-	return (DWORD)sgdwProgress >= 534;
+	return sgdwProgress >= 534;
 }
